Moved booking form dialogue
Now exists within the bookings page, with a 'New Booking' button on the top right
