+++ firebase/src/components/Bookings/BookingPanel.js
@@ -14,7 +14,7 @@ var dateFormat = require('dateformat')
 const useStyles = makeStyles(theme => ({
     heading: {
         fontSize: theme.typography.pxToRem(15),
-        flexBasis: '20%',
+        flexBasis: '40%',
         flexShrink: 0,
     },
     secondaryHeading: {
