+++ firebase/src/constants/routes.js
@@ -1,4 +1,3 @@
 export const LANDING = '/'
 export const SIGN_IN = '/sign-in'
-export const HOME = '/home'
 export const BOOKINGS = '/bookings'
